+++ db/migrate/20140818140042_add_era_commons_username_to_applicants.rb
@@ -1,0 +1,5 @@
+class AddEraCommonsUsernameToApplicants < ActiveRecord::Migration
+  def change
+    add_column :applicants, :era_commons_username, :string
+  end
+end